Reduce the gallery's minimum width
https://gitlab.com/ubports/community-ports/pinephone/-/issues/31 notes
that the gallery goes off the edge of the screen on the PinePhone. This
is due to setting a minimum width, which was likely done to help
developers while designing views.

It does not appear that this large of a minimum is still needed, and it
causes the graphical issues mentioned above on phones that are not very
wide.