Add migration wrapper script
This migrates an existing Qt settings file with the old name if no settings
file using the new name exits yet.
If the pictures directory of lomiri-camera-app has not been renamed yet it will
be renamed.
If pictures in the database still refer to path names inside the pictures
directory of lomiri-camera-app they will be adjusted to the new path as well.